feat: check if there is existing product before updating and deleting

+++ apps/product-service/src/app/product.service.ts
@@ -6,6 +6,7 @@ import { Tokens } from './libs/tokens';
 import { ObjectId } from '@lib/object-id';
 import { FilterQuery } from 'mongoose';
 import { normalizeDocument } from '@lib/util';
+import { ProductError, ProductNotFoundError } from '../error';
 
 @Injectable()
 export class ProductService {
@@ -19,15 +20,27 @@ export class ProductService {
   }
 
   async updateProduct(id: ObjectId, dto: Partial<Omit<Product, 'id'>>) {
+    const existingProduct = await this.findProduct(id);
+
+    if (!existingProduct) {
+      throw new ProductNotFoundError(id.toString());
+    }
     return this.productRepository.update(id, dto);
   }
 
   async deleteProduct(id: ObjectId) {
+    const existingProduct = await this.findProduct(id);
+
+    if (!existingProduct) {
+      throw new ProductNotFoundError(id.toString());
+    }
+
     return this.productRepository.delete(id);
   }
 
   async findProduct(params: FilterQuery<Product>) {
     const product = await this.productRepository.find(params);
+    if (!product) return null;
 
     return normalizeDocument(product);
   }
